# frontend/docs/api-contract.open.yaml
openapi: 3.1.0
info:
  title: DataReaper Frontend Integration API
  version: 1.0.0
  summary: Backend contract inferred from the current DataReaper frontend workspace.
  description: |
    This OpenAPI contract is derived from all active frontend pages and UI state models:
    Landing, Onboarding, Command Center, Identity Graph, and War Room.

    Purpose:
    - Give backend a stable contract to implement for seamless frontend integration.
    - Replace current in-memory/mock UI data with API-driven data.
    - Standardize enums, payloads, pagination, and realtime event shape.

    Notes:
    - Current frontend has no live HTTP integration yet; this is the target contract.
    - Realtime events are defined via HTTP bootstrap endpoint plus extension blocks.
    - All timestamps use ISO 8601 unless explicitly noted.
servers:
  - url: https://api.datareaper.io
    description: Production
  - url: https://staging-api.datareaper.io
    description: Staging
  - url: http://localhost:8080
    description: Local development
...
security:
  - SessionHeader: []
paths:
...
  /v1/scans/{scanId}:
    get:
      tags: [Scans]
      summary: Get scan status and high-level lifecycle state
      operationId: getScan
      parameters:
        - $ref: '#/components/parameters/ScanId'
      responses:
        '200':
          description: Scan status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scan'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /v1/realtime/connection:
    post:
      tags: [Realtime]
      summary: Create realtime connection descriptor (WebSocket or SSE)
      description: |
        Frontend calls this once per scan route mount to receive a short-lived token and endpoint.
        Backend can serve ws://, wss://, or SSE endpoint.
      operationId: createRealtimeConnection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRealtimeConnectionRequest'
      responses:
        '201':
          description: Realtime connection descriptor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRealtimeConnectionResponse'
components:
  securitySchemes:
    SessionHeader:
      type: apiKey
      in: header
      name: X-Session-Id
      description: Anonymous session key from POST /v1/sessions.
  parameters:
    ScanId:
      name: scanId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ScanId'
...
  responses:
...
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
    ScanId:
      type: string
      pattern: '^scn_[A-Za-z0-9]{16,}$'
      example: scn_01JY8H3R8Y9F8S3TB3D3S9K3A1
...
    IsoTimestamp:
      type: string
      format: date-time
      example: '2026-04-18T14:52:31Z'
...
    SeedType:
      type: string
      enum: [email, phone]
    ScanStatus:
      type: string
      enum: [queued, discovering, identifying, engaging, stabilizing, completed, failed]
...
    SeedInput:
      type: object
      required: [type, value]
      properties:
        type:
          $ref: '#/components/schemas/SeedType'
        value:
          type: string
          description: Raw email or phone value from onboarding input.
          example: user@email.com
        countryCode:
          type: string
          description: Required for ambiguous phone parsing.
          example: IN
...
    Scan:
      type: object
      required: [scanId, seed, status, progress, createdAt, updatedAt]
      properties:
        scanId:
          $ref: '#/components/schemas/ScanId'
        seed:
          $ref: '#/components/schemas/SeedInput'
        status:
          $ref: '#/components/schemas/ScanStatus'
        progress:
          type: integer
          minimum: 0
          maximum: 100
        createdAt:
          $ref: '#/components/schemas/IsoTimestamp'
        updatedAt:
          $ref: '#/components/schemas/IsoTimestamp'
        failureReason:
          type: string
...
    CreateRealtimeConnectionRequest:
      type: object
      required: [scanId, channels]
      properties:
        scanId:
          $ref: '#/components/schemas/ScanId'
        channels:
          type: array
          items:
            type: string
            enum:
              - dashboard.summary
              - dashboard.radar
              - dashboard.activity
              - dashboard.agents
              - identity.graph
              - warroom.engagements
              - warroom.messages
              - scans.lifecycle
        preferredTransport:
          type: string
          enum: [websocket, sse]
          default: websocket
    CreateRealtimeConnectionResponse:
      type: object
      required: [connectionId, transport, endpoint, token, expiresAt]
      properties:
        connectionId:
          type: string
          example: rtc_01JY8N3PZG6C7W76P1Z6D4QH0R
        transport:
          type: string
          enum: [websocket, sse]
        endpoint:
          type: string
          format: uri
          example: wss://api.datareaper.io/v1/realtime/ws
        token:
          type: string
          description: Short-lived bearer used only for realtime handshake.
        expiresAt:
          $ref: '#/components/schemas/IsoTimestamp'
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: invalid_seed
        message:
          type: string
          example: Seed value must be a valid email or E.164 phone number.
        details:
          type: array
          items:
            type: object
            additionalProperties: true
...